Move HTML formatting up in the display process
Move HTML formatting and linkification out of the hyperlink delegate and into its own utils file, then format output text at the original display model instead of in each handler that interacts with the text independently.
